requirements.txt
apiclient==1.0.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
gspread==3.1.0
gspread-dataframe==3.0.3
httplib2==0.15.0
idna==2.8
line-bot-sdk==1.15.0
numpy==1.18.1
oauth2client==4.1.3
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.13.0
uritemplate==3.0.1
urllib3==1.25.7
